% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{Covid19EstimationHydroxychloroquine}
\alias{Covid19EstimationHydroxychloroquine}
\alias{Covid19EstimationHydroxychloroquine-package}
\title{Covid19EstimationHydroxychloroquine}
\description{
Covid19EstimationHydroxychloroquine
}
